refactor(frontend): extract routes, ProfileRedirect, LayoutProviders, useLogout, Desktop/* from App/Layout
Splits App.tsx (178 lines) and Layout.tsx (284 lines) into focused files
with zero behavior change: mounted component tree, provider order/nesting,
route paths, lazy/eager split, and all props are unchanged.

- routes.tsx: the 23 lazy + 5 eager route imports and the <Routes> tree
  (App.tsx keeps Suspense/ErrorBoundary/RouteErrorBoundary wrapping it)
- components/ProfileRedirect.tsx: the bare `/profile` redirect, moved
  verbatim, loose at components/ root (matches AuthGate/PublicRoute)
- components/LayoutProviders.tsx: the ReplayBufferProvider >
  SocketHubProvider > IncomingCallProvider stack that was triplicated
  across Layout's mobile/tablet/desktop branches, plus LayoutHooksBridge
  (now private), IncomingCallListener, IncomingCallBanner
- hooks/useLogout.ts: handleLogout + logoutLoading, preserving the exact
  operation order (leave voice -> clearSavedConnection ->
  disconnectSocket -> Electron refresh-token branch -> logout mutation ->
  clearTokens -> navigate)
- components/Desktop/{DesktopLayout,DesktopAppBar,DesktopContentArea}.tsx
  + barrel: the desktop branch body, AppBar/Toolbar chrome (now owning
  useLogout + the user menu), and the renamed LayoutContentArea

Known wart preserved: useVoiceRecovery is still double-invoked on mobile
(Layout + MobileLayout) — left as-is with a TODO comment, per the
existing double-call pattern already present for useVoiceConnection on
mobile/tablet.
