+++ frontend/src/components/Desktop/index.ts
@@ -1,0 +1,9 @@
+/**
+ * Desktop Layout Components
+ *
+ * Full AppBar + sidebar + voice bottom bar layout for desktop (>= 1200px).
+ */
+
+export { DesktopLayout } from './DesktopLayout';
+export { DesktopAppBar } from './DesktopAppBar';
+export { DesktopContentArea } from './DesktopContentArea';
